Move unused CompareDist to attic
