+++ timemachine/fe/atom_mapping.py
@@ -7,24 +7,6 @@ from rdkit.Chem import rdFMCS
 from timemachine.fe.topology import AtomMappingError
 
 
-class CompareDist(rdFMCS.MCSAtomCompare):
-    """Custom atom comparison: use positions within generated conformer"""
-
-    def __init__(self, threshold=0.5, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.threshold = threshold
-
-    def compare(self, p, mol1, atom1, mol2, atom2):
-        """Atoms match if within 0.5 Å
-
-        Signature from super method:
-        (MCSAtomCompareParameters)parameters, (Mol)mol1, (int)atom1, (Mol)mol2, (int)atom2) -> bool
-        """
-        x_i = mol1.GetConformer(0).GetPositions()[atom1]
-        x_j = mol2.GetConformer(0).GetPositions()[atom2]
-        return bool(np.linalg.norm(x_i - x_j) <= self.threshold)  # must convert from np.bool_ to Python bool!
-
-
 class CompareDistNonterminal(rdFMCS.MCSAtomCompare):
     """
     Custom comparator used in the FMCS code.
